Prevent unsound uses of DerivePConstantViaData
